+++ assignments/02-sdh/drone-hangar/src/core/timer.cpp
@@ -5,10 +5,10 @@
 
 #include "config.h"
 
-Timer::Timer(uint64_t periodMillis) : periodMillis(periodMillis), startTime(0) {}
+Timer::Timer(uint32_t periodMillis) : periodMillis(periodMillis), startTime(0) {}
 Timer::Timer() : Timer(0) {}
 
-uint64_t Timer::getElapsedMillis() { return millis() - this->startTime; }
+uint32_t Timer::getElapsedMillis() { return millis() - this->startTime; }
 
 bool Timer::isFinished()
 {
@@ -20,12 +20,12 @@ void Timer::start()
   this->startTime = millis();
 }
 
-uint64_t Timer::wait()
+uint32_t Timer::wait()
 {
   while (this->getElapsedMillis() < this->periodMillis)
   {
     delay(TIMER_READ_FREQ_MS);
   }
-  uint64_t total = this->getElapsedMillis();
+  uint32_t total = this->getElapsedMillis();
   return total;
 }
