+++ assignments/02-sdh/drone-hangar/src/devices/sonar.cpp
@@ -4,8 +4,8 @@
 #include <limits.h>
 
 UltrasonicSensor::UltrasonicSensor(uint8_t echoPin, uint8_t triggerPin,
-                                   uint64_t readStartMicros, uint64_t readDelayMicros,
-                                   uint64_t readTimeoutMicros, TemperatureSensor *tempSensor,
+                                   uint32_t readStartMicros, uint32_t readDelayMicros,
+                                   uint32_t readTimeoutMicros, TemperatureSensor *tempSensor,
                                    float minDistanceMm, float maxDistanceMm)
     : echoPin(echoPin),
       triggerPin(triggerPin),
@@ -31,7 +31,7 @@ float UltrasonicSensor::getSoundSpeed()
 }
 
 // https://github.com/pslab-unibo/esiot-2025-2026/blob/18463be0de31a90caf9740de2df1c6108cb350b3/lab-activities/lab-activity-08/sweeping-system/arduino/sweeping-system/src/devices/Sonar.cpp#L29-L30
-float UltrasonicSensor::pulseToDistance(uint64_t pulse)
+float UltrasonicSensor::pulseToDistance(uint32_t pulse)
 {
   return ((float)pulse) / 1000.0 / 2 * this->getSoundSpeed(); // one less x/1000 to get mm
 }
@@ -44,10 +44,10 @@ void UltrasonicSensor::read()
   delayMicroseconds(this->readDelayMicros);
   this->triggerPin.write(DigitalValue::Low);
 
-  uint64_t readStart = micros();
-  uint64_t pulse =
+  uint32_t readStart = micros();
+  uint32_t pulse =
       this->echoPin.readPulse(DigitalValue::High, this->readTimeoutMicros);
-  uint64_t readTime = micros() - readStart;
+  uint32_t readTime = micros() - readStart;
   if (pulse == 0)
   {
     this->distanceMm = readTime < this->readTimeoutMicros ? this->minDistanceMm : this->maxDistanceMm;
