+++ assignments/02-sdh/drone-hangar/src/tasks/alarm.cpp
@@ -12,8 +12,8 @@ AlarmTask::AlarmTempCheckingState AlarmTask::ALARM_TEMP_CHECKING;
 
 AlarmTask::AlarmTask(TemperatureSensor *hangarTempSensor,
                      Indicator *alarmIndicator, GlobalState initialState,
-                     float prealarmTemp, uint64_t prealarmTimeMillis,
-                     float alarmTemp, uint64_t alarmTimeMillis)
+                     float prealarmTemp, uint32_t prealarmTimeMillis,
+                     float alarmTemp, uint32_t alarmTimeMillis)
     : Task<AlarmTask>(&AlarmTask::IDLE),
       hangarTempSensor(hangarTempSensor),
       alarmIndicator(alarmIndicator),
