+++ assignments/02-sdh/drone-hangar/src/tasks/blink.hpp
@@ -7,7 +7,7 @@ class BlinkTask : public Task<BlinkTask>
 {
 private:
   Indicator *blinkIndicator;
-  uint64_t periodMillis;
+  uint32_t periodMillis;
   Timer timer;
 
   class IdleState : public TaskState<BlinkTask>
@@ -34,5 +34,5 @@ private:
   static OffState OFF;
 
 public:
-  BlinkTask(Indicator *blinkIndicator, uint64_t periodMillis);
+  BlinkTask(Indicator *blinkIndicator, uint32_t periodMillis);
 };
